+++ common/Makefile
@@ -26,6 +26,7 @@ COMMON_SRC_NOGEN :=				\
 	common/io_lock.c			\
 	common/json.c				\
 	common/json_escaped.c			\
+	common/json_helpers.c			\
 	common/json_tok.c			\
 	common/key_derive.c			\
 	common/keyset.c				\
