+++ common/json.c
@@ -1,6 +1,7 @@
 /* JSON core and helpers */
 #include "json.h"
 #include <assert.h>
+#include <bitcoin/pubkey.h>
 #include <ccan/build_assert/build_assert.h>
 #include <ccan/mem/mem.h>
 #include <ccan/str/hex/hex.h>
@@ -125,34 +126,6 @@ bool json_to_bool(const char *buffer, const jsmntok_t *tok, bool *b)
 		return true;
 	}
 	return false;
-}
-
-bool json_to_bitcoin_amount(const char *buffer, const jsmntok_t *tok,
-			    uint64_t *satoshi)
-{
-	char *end;
-	unsigned long btc, sat;
-
-	btc = strtoul(buffer + tok->start, &end, 10);
-	if (btc == ULONG_MAX && errno == ERANGE)
-		return false;
-	if (end != buffer + tok->end) {
-		/* Expect always 8 decimal places. */
-		if (*end != '.' || buffer + tok->end - end != 9)
-			return false;
-		sat = strtoul(end+1, &end, 10);
-		if (sat == ULONG_MAX && errno == ERANGE)
-			return false;
-		if (end != buffer + tok->end)
-			return false;
-	} else
-		sat = 0;
-
-	*satoshi = btc * (uint64_t)100000000 + sat;
-	if (*satoshi != btc * (uint64_t)100000000 + sat)
-		return false;
-
-	return true;
 }
 
 bool json_tok_is_num(const char *buffer, const jsmntok_t *tok)
