+++ lightningd/gossip_control.c
@@ -13,6 +13,7 @@
 #include <common/features.h>
 #include <common/json_command.h>
 #include <common/json_escaped.h>
+#include <common/json_helpers.h>
 #include <common/jsonrpc_errors.h>
 #include <common/param.h>
 #include <common/type_to_string.h>
