+++ lightningd/peer_control.c
@@ -16,6 +16,7 @@
 #include <common/features.h>
 #include <common/initial_commit_tx.h>
 #include <common/json_command.h>
+#include <common/json_helpers.h>
 #include <common/jsonrpc_errors.h>
 #include <common/key_derive.h>
 #include <common/param.h>
